# src/main/java/io/github/sudoitir/artemisstudio/platform/clusters/web/ClusterController.java
package io.github.sudoitir.artemisstudio.platform.clusters.web;

import io.github.sudoitir.artemisstudio.kernel.core.Attempt;
import io.github.sudoitir.artemisstudio.platform.broker.BrokerConnectionException;
import io.github.sudoitir.artemisstudio.platform.clusters.ClusterService;
import io.github.sudoitir.artemisstudio.platform.clusters.web.ClusterRequests.NodeOverrideRequest;
import io.github.sudoitir.artemisstudio.platform.clusters.web.ClusterRequests.RegisterClusterRequest;
import io.github.sudoitir.artemisstudio.platform.clusters.web.ClusterRequests.RotateCredentialsRequest;
import io.github.sudoitir.artemisstudio.platform.clusters.web.ClusterViews.CapabilitiesView;
import io.github.sudoitir.artemisstudio.platform.clusters.web.ClusterViews.ClusterDetail;
import io.github.sudoitir.artemisstudio.platform.clusters.web.ClusterViews.ClusterSummary;
import io.github.sudoitir.artemisstudio.platform.clusters.web.ClusterViews.HealthView;
import io.github.sudoitir.artemisstudio.platform.clusters.web.ClusterViews.NodeEndpointView;
import io.github.sudoitir.artemisstudio.platform.clusters.web.ClusterViews.RegisterPreview;
import io.github.sudoitir.artemisstudio.platform.clusters.web.ClusterViews.TopologyView;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import jakarta.validation.Valid;
import java.util.List;
import java.util.UUID;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

/**
 * The Phase 1 cluster API (base {@code /api/v1}). Realtime is SSE only (ADR-0003);
 * every command here is an ordinary POST/PATCH/DELETE. Destructive operations take
 * {@code ?dryRun=true} and typed confirmation is enforced in the UI
 * (non-negotiable #2).
 */
@RestController
@RequestMapping("/api/v1/clusters")
@RequiredArgsConstructor
public class ClusterController {

    private final ClusterService service;

    /** Register from a list of seed URLs (ADR-0013). {@code ?dryRun=true} probes and returns without persisting. */
    @ApiResponse(
            responseCode = "200",
            description = "dryRun=true: connection probe result, nothing persisted",
            content = @Content(schema = @Schema(implementation = RegisterPreview.class)))
    @ApiResponse(
            responseCode = "201",
            description = "cluster registered",
            content = @Content(schema = @Schema(implementation = ClusterDetail.class)))
    @PostMapping
    public ResponseEntity<Object> register(
            @Valid @RequestBody RegisterClusterRequest request, @RequestParam(defaultValue = "false") boolean dryRun) {
        if (dryRun) {
            return ResponseEntity.ok(unwrap(service.checkConnection(request)));
        }
        ClusterDetail detail = unwrap(service.register(request));
        return ResponseEntity.status(HttpStatus.CREATED).body(detail);
    }

    @GetMapping
    public List<ClusterSummary> list() {
        return service.list();
    }

    @GetMapping("/{clusterId}")
    public ClusterDetail get(@PathVariable UUID clusterId) {
        return service.get(clusterId);
    }

    @DeleteMapping("/{clusterId}")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void delete(@PathVariable UUID clusterId) {
        service.delete(clusterId);
    }

    @GetMapping("/{clusterId}/capabilities")
    public CapabilitiesView capabilities(@PathVariable UUID clusterId) {
        return service.capabilities(clusterId);
    }

    @GetMapping("/{clusterId}/topology")
    public TopologyView topology(@PathVariable UUID clusterId) {
        return service.topology(clusterId);
    }

    @GetMapping("/{clusterId}/health")
    public HealthView health(@PathVariable UUID clusterId) {
        return service.health(clusterId);
    }

    @PostMapping("/{clusterId}/rediscover")
    public TopologyView rediscover(@PathVariable UUID clusterId) {
        return unwrap(service.rediscover(clusterId));
    }

    @PutMapping("/{clusterId}/credentials")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void rotateCredentials(@PathVariable UUID clusterId, @Valid @RequestBody RotateCredentialsRequest request) {
        service.rotateCredentials(clusterId, request.username(), request.password(), request.kindOrDefault());
    }

    @PatchMapping("/{clusterId}/nodes/{nodeId}")
    public NodeEndpointView overrideNode(
            @PathVariable UUID clusterId, @PathVariable UUID nodeId, @Valid @RequestBody NodeOverrideRequest request) {
        return unwrap(service.overrideNodeUrl(clusterId, nodeId, request));
    }

    /** A {@link Attempt.Failed} becomes a classified {@link BrokerConnectionException} for the advice to render. */
    private static <T> T unwrap(Attempt<T> attempt) {
        return switch (attempt) {
            case Attempt.Ok<T> ok -> ok.value();
            case Attempt.Failed<T> failed -> throw new BrokerConnectionException(failed.kind(), failed.detail());
        };
    }
}