# web/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Artemis Studio API",
    "description" : "Cluster-wide management and observability for Apache ActiveMQ Artemis",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/clusters/{clusterId}/credentials" : {
      "put" : {
        "tags" : [ "cluster-controller" ],
        "operationId" : "rotateCredentials",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RotateCredentialsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/v1/clusters" : {
      "get" : {
        "tags" : [ "cluster-controller" ],
        "operationId" : "list_5",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ClusterSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "cluster-controller" ],
        "operationId" : "register",
        "parameters" : [ {
          "name" : "dryRun",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterClusterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "dryRun=true: connection probe result, nothing persisted",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RegisterPreview"
                }
              }
            }
          },
          "201" : {
            "description" : "cluster registered",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClusterDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}/rediscover" : {
      "post" : {
        "tags" : [ "cluster-controller" ],
        "operationId" : "rediscover",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TopologyView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}/nodes/{nodeId}" : {
      "patch" : {
        "tags" : [ "cluster-controller" ],
        "operationId" : "overrideNode",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "nodeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NodeOverrideRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NodeEndpointView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}" : {
      "get" : {
        "tags" : [ "cluster-controller" ],
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClusterDetail"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "cluster-controller" ],
        "operationId" : "delete_4",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/v1/clusters/{clusterId}/topology" : {
      "get" : {
        "tags" : [ "cluster-controller" ],
        "operationId" : "topology",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TopologyView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clusters/{clusterId}/capabilities" : {
      "get" : {
        "tags" : [ "cluster-controller" ],
        "operationId" : "capabilities",
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CapabilitiesView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "RotateCredentialsRequest" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string",
            "minLength" : 1
          },
          "password" : {
            "type" : "string",
            "minLength" : 1
          },
          "kind" : {
            "type" : "string"
          }
        },
        "required" : [ "password", "username" ]
      },
...
      "Credentials" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string",
            "minLength" : 1
          },
          "password" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "password", "username" ]
      },
      "RegisterClusterRequest" : {
        "type" : "object",
        "properties" : {
          "seedUrls" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "minLength" : 1
            },
            "minItems" : 1
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "credentials" : {
            "$ref" : "#/components/schemas/Credentials"
          },
          "coreCredentials" : {
            "$ref" : "#/components/schemas/Credentials"
          },
          "tlsBundle" : {
            "type" : "string"
          }
        },
        "required" : [ "seedUrls" ]
      },
      "CapabilitiesView" : {
        "type" : "object",
        "properties" : {
          "managementRead" : {
            "$ref" : "#/components/schemas/CapabilityView"
          },
          "managementWrite" : {
            "$ref" : "#/components/schemas/CapabilityView"
          },
          "notifications" : {
            "$ref" : "#/components/schemas/CapabilityView"
          },
          "messageIo" : {
            "$ref" : "#/components/schemas/CapabilityView"
          },
          "slowConsumerDetection" : {
            "$ref" : "#/components/schemas/CapabilityView"
          }
        },
        "required" : [ "managementRead", "managementWrite", "messageIo", "notifications", "slowConsumerDetection" ]
      },
      "CapabilityView" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "brokerXmlSnippet" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "reason", "status" ]
      },
      "ConfigRecommendationView" : {
        "type" : "object",
        "description" : "One capability gap and what would close it",
        "properties" : {
          "capability" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "rationale" : {
            "type" : "string"
          },
          "appliable" : {
            "type" : "boolean",
            "description" : "Whether Studio can write this over the management API, or the operator must edit broker.xml and restart"
          },
          "section" : {
            "type" : [ "string", "null" ],
            "enum" : [ "ADDRESS_SETTING", "SECURITY_SETTING" ]
          },
          "match" : {
            "type" : [ "string", "null" ]
          },
          "values" : {
            "type" : "object",
            "additionalProperties" : { },
            "description" : "The whole entry that would be written, the node's current keys included — a runtime write replaces the entry rather than merging"
          },
          "roles" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            },
            "description" : "Permission type to role names, prefilled from the broker"
          },
          "keys" : {
            "type" : "array",
            "description" : "The keys this recommendation itself sets",
            "items" : {
              "type" : "string"
            }
          },
          "manualSnippet" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "appliable", "capability", "keys", "rationale", "roles", "title", "values" ]
      },
      "ConfigRecommendationsView" : {
        "type" : "object",
        "description" : "What the capability probe suggests declaring, and what still needs a broker.xml edit",
        "properties" : {
          "seededFrom" : {
            "type" : [ "string", "null" ],
            "description" : "The node the current values were read from; null when none could be read"
          },
          "recommendations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ConfigRecommendationView"
            }
          }
        },
        "required" : [ "recommendations" ]
      },
      "LogicalNodeView" : {
        "type" : "object",
        "properties" : {
          "artemisNodeId" : {
            "type" : [ "string", "null" ]
          },
          "splitBrain" : {
            "type" : "string"
          },
          "replicationBehind" : {
            "type" : "boolean"
          },
          "endpoints" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/NodeEndpointView"
            }
          }
        },
        "required" : [ "endpoints", "replicationBehind", "splitBrain" ]
      },
      "NodeEndpointView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "artemisNodeId" : {
            "type" : [ "string", "null" ]
          },
          "jolokiaUrl" : {
            "type" : [ "string", "null" ]
          },
          "coreUrl" : {
            "type" : [ "string", "null" ]
          },
          "haRole" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string"
          },
          "active" : {
            "type" : "boolean"
          },
          "replicaSync" : {
            "type" : [ "boolean", "null" ]
          },
          "version" : {
            "type" : [ "string", "null" ]
          },
          "lastError" : {
            "type" : [ "string", "null" ]
          },
          "lastSeenAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "discovered" : {
            "type" : "boolean"
          },
          "manualOverride" : {
            "type" : "boolean"
          },
          "manageable" : {
            "type" : "boolean"
          }
        },
        "required" : [ "active", "discovered", "haRole", "id", "manageable", "manualOverride", "name", "state" ]
      },
      "RegisterPreview" : {
        "type" : "object",
        "properties" : {
          "capabilities" : {
            "$ref" : "#/components/schemas/CapabilitiesView"
          },
          "reachableSeeds" : {
            "type" : "integer",
            "format" : "int32"
          },
          "discoveredNodes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "topology" : {
            "$ref" : "#/components/schemas/TopologyView"
          },
          "recommendations" : {
            "$ref" : "#/components/schemas/ConfigRecommendationsView",
            "description" : "What the probe suggests declaring once the cluster is registered, seeded from what the reachable node is running"
          }
        },
        "required" : [ "capabilities", "discoveredNodes", "reachableSeeds", "recommendations", "topology" ]
      },
      "TopologyView" : {
        "type" : "object",
        "properties" : {
          "clusterId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "nodes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LogicalNodeView"
            }
          }
        },
        "required" : [ "clusterId", "nodes" ]
      },
      "ClusterDetail" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "topology" : {
            "$ref" : "#/components/schemas/TopologyView"
          },
          "capabilities" : {
            "$ref" : "#/components/schemas/CapabilitiesView"
          },
          "health" : {
            "$ref" : "#/components/schemas/HealthView"
          },
          "environmentId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          }
        },
        "required" : [ "capabilities", "health", "id", "name", "topology" ]
      },
      "HealthView" : {
        "type" : "object",
        "properties" : {
          "clusterId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "level" : {
            "type" : "string"
          },
          "liveEndpointNames" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "splitBrain" : {
            "type" : "string"
          },
          "replicationBehind" : {
            "type" : "boolean"
          },
          "notes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "clusterId", "level", "liveEndpointNames", "notes", "replicationBehind", "splitBrain" ]
      },
...
      "NodeOverrideRequest" : {
        "type" : "object",
        "properties" : {
          "jolokiaUrl" : {
            "type" : "string"
          },
          "coreUrl" : {
            "type" : "string"
          }
        }
      },
...
      "ClusterSummary" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "health" : {
            "type" : "string",
            "enum" : [ "OK", "DEGRADED", "CRITICAL", "UNKNOWN" ]
          },
          "nodeCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "environmentId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          }
        },
        "required" : [ "health", "id", "name", "nodeCount", "updatedAt" ]
      },
...
    }
  }
}